pyproject.toml
[project]
name = "twilio-task"
version = "0.1.0"
description = ""
authors = [
    {name = "qw1ndy",email = "kitsor05@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "python-dotenv (>=1.1.1,<2.0.0)",
    "fastapi (>=0.115.14,<0.116.0)",
    "twilio (>=9.6.4,<10.0.0)",
    "google (>=3.0.0,<4.0.0)",
    "google-genai (>=1.24.0,<2.0.0)",
    "asyncio (>=3.4.3,<4.0.0)",
    "websockets (>=15.0.1,<16.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "langgraph (>=0.5.2,<0.6.0)",
    "langchain (>=0.3.26,<0.4.0)",
    "langchain-google-genai (>=2.1.7,<3.0.0)"
]

[tool.poetry]
packages = [{include = "twilio_task", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
